Simplify Ast
It is better to use strings in the parser, instead of constructing
arrays for words. It uses less memory for a certaintey and probably
performs better in time also. Given that strings in Scala are character
sequences, strings are just as easily handled as any other collection
when iterating over them.