[diko] Occupation rate configurable
The generated compiled Fst is compressed so that it does not take too
much memory on disk. Compression can take time due to the used
"first-fit" algorithm to store states. Passed a given occupation rate,
the available holes are not sufficient to make new tates fit into them,
and most of the time is used to try position from the beginning to the
end before deciding we must save the state at the end.

Measure showed that on test sample 84% occupation is achieved. However
if we decide to start occupying a new fresh sector as soon as 70% of the
old sector is used, then we can achieve 69% occupation by deviding time
needed to compute by a factor of 6 and by still having a decent
occupation rate. This limit is the new default, but it can be tweaked
using options.